+++ CHANGES.md
@@ -1,3 +1,9 @@
+## v0.6.0 
+
+- registry: Add support for a Parquet registry as an alternative source
+  to lookup tile information (#16).
+- docs: Fix old documentation examples (#22, #18 report by @cjissmart)
+
 ## v0.5.2 (2025-09-02)
 
 - cli: Add date/time/repo/hash information to the coverage maps